refactor(gg_api_core): move client identity and tool tracking into dedicated modules
Responds to review on the MCP client-visibility headers:

- Split client identity out of log_context.py into client_identity.py, keyed
  by the session in a WeakKeyDictionary instead of setattr-ing the identity
  onto FastMCP's session object (no foreign-object mutation, auto-GC).
  set_client_identity now types session as ServerSession.
- Split the current-tool tracking (track_current_tool / current_tool_name)
  out of log_context.py into tool_context.py, per the review request to stop
  keeping those globals in the middle of log_context.
- Update middleware, client, and utils imports to the new modules; drop the
  comments the review flagged.
- Remove the overkill TestClientIdentity block from test_log_context.py; test
  fakes now subclass ServerSession so they satisfy the typed signature.
