+++ tests/test_client.py
@@ -4,7 +4,7 @@ from unittest.mock import AsyncMock, MagicMock, patch
 import httpx
 import pytest
 from gg_api_core.client import GitGuardianClient, IncidentSeverity, IncidentStatus, IncidentValidity
-from gg_api_core.log_context import track_current_tool
+from gg_api_core.tool_context import track_current_tool
 
 
 @pytest.fixture
